feat: add FilledButton component for testing
